Visitor counting project using a laser and diode.
This project is developed for Thinks Labs. This project includes a
embedded system to count the number of visitors in a public place. This
device helps in emergency and disaster management situations. It
includes a laser and diode pair to count the visitors at the entry
point. Software developed takes into consideration lots of
possibilities, maximum count and display functionalities.